fix(list): fix builder mixing up order for larger lists
fix #46
