+++ deno_dist/list/builder/tree/tree-builder.ts
@@ -1,16 +1,9 @@
 import { OptLazy, TraverseState, Update } from '../../../common/mod.ts';
 import type {
   BlockBuilder,
-  BuilderBase,
-  LeafBlockBuilder,
-  LeafBuilder,
-  LeafTree,
   ListContext,
-  NonLeafBlockBuilder,
   NonLeafBuilder,
-  NonLeafTree,
 } from '../../list-custom.ts';
-import { createFromBlock, createNonLeaf } from '../../list-custom.ts';
 
 export abstract class TreeBuilderBase<T, C> {
   abstract get context(): ListContext;
@@ -242,7 +235,9 @@ export abstract class TreeBuilderBase<T, C> {
           }
         );
 
-        if (undefined !== delta) return;
+        if (undefined !== delta) {
+          return;
+        }
       } else if (this.right.nrChildren < this.context.maxBlockSize) {
         // try to shift to right
         const shiftChild = this.left.dropLast();
@@ -458,269 +453,5 @@ export abstract class TreeBuilderBase<T, C> {
     }
 
     this.right.forEach(f, state);
-  }
-}
-
-export class LeafTreeBuilder<T>
-  extends TreeBuilderBase<T, T>
-  implements LeafBuilder<T>
-{
-  constructor(
-    readonly context: ListContext,
-    public source?: LeafTree<T>,
-    public _left?: LeafBlockBuilder<T>,
-    public _right?: LeafBlockBuilder<T>,
-    public _middle?: NonLeafBuilder<T, LeafBlockBuilder<T>>,
-    public length = source?.length ?? 0
-  ) {
-    super();
-  }
-
-  prepareMutate(): void {
-    if (undefined !== this.source) {
-      this._left = this.context.leafBlockBuilderSource(this.source.left);
-      this._right = this.context.leafBlockBuilderSource(this.source.right);
-      this._middle =
-        null === this.source.middle
-          ? undefined
-          : (createNonLeaf<T>(this.source.middle) as any);
-      this.source = undefined;
-    }
-  }
-
-  get level(): 0 {
-    return 0;
-  }
-
-  get left(): LeafBlockBuilder<T> {
-    this.prepareMutate();
-    return this._left!;
-  }
-
-  set left(value: LeafBlockBuilder<T>) {
-    this.prepareMutate();
-    this._left = value;
-  }
-
-  get right(): LeafBlockBuilder<T> {
-    this.prepareMutate();
-    return this._right!;
-  }
-
-  set right(value: LeafBlockBuilder<T>) {
-    this.prepareMutate();
-    this._right = value;
-  }
-
-  get middle(): NonLeafBuilder<T, LeafBlockBuilder<T>> | undefined {
-    this.prepareMutate();
-    return this._middle;
-  }
-
-  set middle(value: NonLeafBuilder<T, LeafBlockBuilder<T>> | undefined) {
-    this.prepareMutate();
-    this._middle = value;
-  }
-
-  getChildLength(): 1 {
-    return 1;
-  }
-
-  appendChildren(children: T[], from: number): void {
-    if (children.length === 0 || from >= children.length) return;
-
-    if (this.right.nrChildren < this.context.maxBlockSize) {
-      const items = children.slice(
-        from,
-        from + this.context.maxBlockSize - this.right.nrChildren
-      );
-      this.right.children = this.right.children.concat(items);
-      this.length += items.length;
-      this.appendChildren(children, from + items.length);
-      return;
-    }
-
-    const block = children.slice(from, from + this.context.maxBlockSize);
-    this.appendMiddle(this.right);
-    this.right = this.context.leafBlockBuilder(block);
-    this.length += block.length;
-    this.appendChildren(children, from + block.length);
-  }
-
-  normalized(): LeafBuilder<T> {
-    if (this.length <= this.context.maxBlockSize) {
-      this.left.concat(this.right);
-      return this.left;
-    }
-
-    if (undefined !== this.middle) {
-      if (this.middle.length + this.left.length <= this.context.maxBlockSize) {
-        this.left.concat(this.middle.first());
-        this.middle = undefined;
-      } else if (
-        this.middle.length + this.right.length <=
-        this.context.maxBlockSize
-      ) {
-        const newRight = this.middle.last();
-        newRight.concat(this.right);
-        this.right = newRight;
-        this.middle = undefined;
-      }
-    }
-
-    return this;
-  }
-
-  get<O>(index: number, otherwise?: OptLazy<O>): T | O {
-    if (undefined !== this.source) return this.source.get(index, otherwise);
-
-    return super.get(index, otherwise);
-  }
-
-  build(): LeafTree<T> {
-    if (undefined !== this.source) return this.source;
-
-    return this.context.leafTree(
-      this.left.build(),
-      this.right.build(),
-      undefined === this.middle ? null : this.middle.build()
-    );
-  }
-
-  buildMap<T2>(f: (value: T) => T2): LeafTree<T2> {
-    if (undefined !== this.source) return this.source.map(f);
-
-    return this.context.leafTree(
-      this.left.buildMap(f),
-      this.right.buildMap(f),
-      undefined === this.middle ? null : (this.middle as any).buildMap(f)
-    );
-  }
-}
-
-export class NonLeafTreeBuilder<T, C extends BlockBuilder<T>>
-  extends TreeBuilderBase<T, C>
-  implements BuilderBase<T, C>
-{
-  constructor(
-    readonly context: ListContext,
-    readonly level: number,
-    public source?: NonLeafTree<T, any>,
-    public _left?: NonLeafBlockBuilder<T, C>,
-    public _right?: NonLeafBlockBuilder<T, C>,
-    public _middle?: NonLeafBuilder<T, NonLeafBlockBuilder<T, C>>,
-    public length = source?.length ?? 0
-  ) {
-    super();
-  }
-
-  prepareMutate(): void {
-    if (undefined !== this.source) {
-      this._left = createFromBlock(this.source.left) as any;
-      this._right = createFromBlock(this.source.right) as any;
-      this._middle =
-        null === this.source.middle
-          ? undefined
-          : (createNonLeaf(this.source.middle) as any);
-      this.source = undefined;
-    }
-  }
-
-  get left(): NonLeafBlockBuilder<T, C> {
-    this.prepareMutate();
-    return this._left!;
-  }
-
-  set left(value: NonLeafBlockBuilder<T, C>) {
-    this.prepareMutate();
-    this._left = value;
-  }
-
-  get right(): NonLeafBlockBuilder<T, C> {
-    this.prepareMutate();
-    return this._right!;
-  }
-
-  set right(value: NonLeafBlockBuilder<T, C>) {
-    this.prepareMutate();
-    this._right = value;
-  }
-
-  get middle(): NonLeafBuilder<T, NonLeafBlockBuilder<T, C>> | undefined {
-    this.prepareMutate();
-    return this._middle;
-  }
-
-  set middle(value: NonLeafBuilder<T, NonLeafBlockBuilder<T, C>> | undefined) {
-    this.prepareMutate();
-    this._middle = value;
-  }
-
-  getChildLength(child: C): number {
-    return child.length;
-  }
-
-  modifyFirstChild(f: (child: C) => number | undefined): number | undefined {
-    const delta = this.left.modifyFirstChild(f);
-    if (undefined !== delta) this.length += delta;
-    return delta;
-  }
-
-  modifyLastChild(f: (child: C) => number | undefined): number | undefined {
-    const delta = this.right.modifyLastChild(f);
-    if (undefined !== delta) this.length += delta;
-    return delta;
-  }
-
-  normalized(): NonLeafBuilder<T, C> {
-    if (undefined === this.middle) {
-      if (
-        this.left.nrChildren + this.right.nrChildren <=
-        this.context.maxBlockSize
-      ) {
-        this.left.concat(this.right);
-        return this.left;
-      }
-    } else {
-      this.middle = this.middle.normalized();
-    }
-
-    return this;
-  }
-
-  first(): C {
-    return this.left.first();
-  }
-
-  last(): C {
-    return this.right.last();
-  }
-
-  get<O>(index: number, otherwise?: OptLazy<O>): T | O {
-    if (undefined !== this.source) return this.source.get(index);
-
-    return super.get(index, otherwise);
-  }
-
-  build(): NonLeafTree<T, any> {
-    if (undefined !== this.source) return this.source;
-
-    return this.context.nonLeafTree(
-      this.left.build(),
-      this.right.build(),
-      undefined === this.middle ? null : this.middle.build(),
-      this.level
-    );
-  }
-
-  buildMap<T2>(f: (value: T) => T2): NonLeafTree<T2, any> {
-    if (undefined !== this.source) return this.source.map(f);
-
-    return this.context.nonLeafTree(
-      this.left.buildMap(f),
-      this.right.buildMap(f),
-      undefined === this.middle ? null : (this.middle as any).buildMap(f),
-      this.level
-    );
   }
 }
